// packages/core/generated/types/github-workflow.d.ts
export type StringContainingExpressionSyntax = string;
// ...
export type ExpressionSyntax = string;
// ...
export interface NormalJob {
  /**
   * The name of the job displayed on GitHub.
   */
  name?: string;
  needs?: JobNeeds;
  permissions?: Permissions;
  /**
   * The type of machine to run the job on. The machine can be either a GitHub-hosted runner, or a self-hosted runner.
   */
  'runs-on':
    | string
    | ([string, ...string[]] & unknown[])
    | {
        group?: string;
        labels?: string | string[];
      }
    | StringContainingExpressionSyntax
    | ExpressionSyntax;
// ...
}